storage/requirements.txt
connexion==2.14.2
swagger-ui-bundle==0.0.9
SQLAlchemy==1.4.50
mysql-connector-python==8.0.30
pyMySQL==1.0.2
pykafka==2.8.0
